+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Added `toMap`, `toSet`.
+
 # 1.0.0
 
 Initial version.